requirements.txt
google-adk>=1.20.0